Modification de l'architecture et création de la route editrecipe ainsi que son composant.
