+++ frontend/src/pages/user/pages/EditRecipe/EditRecipe.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const EditRecipe = () => {
+  return (
+    <div>EditRecipe</div>
+  )
+}
+
+export default EditRecipe
