+++ frontend/src/pages/user/pages/components/FilteredIngredientsDisplay.jsx
