+++ frontend/src/pages/user/pages/components/selectForm/SelectedIngredientsDisplay.jsx
