+++ frontend/src/pages/user/pages/components/selectForm/TypeSelect.jsx
@@ -1,18 +1,15 @@
 import React from 'react';
-import { faList } from "@fortawesome/free-solid-svg-icons";
-import Select from '../../../../../../components/form/Select'; 
+import Select from '../../../../../components/form/Select';
+import { faList } from '@fortawesome/free-solid-svg-icons';
 
-const CategorySelect = ({ selectedCategory, categories, handleCategoryChange }) => {
-  const options = [
-    { value: '', label: 'Toutes les catégories' },
-    ...categories.map(category => ({
-      value: category._id,
-      label: category.name
-    }))
+const TypeSelect = ({ selectedType, filteredTypes, handleTypeChange, showIngredients }) => {
+  const option = [
+    { value : '', label: 'Tous les types' },
+    ...filteredTypes.map(type => ({ value: type._id, label: type.name })),
   ];
 
   const handleChange = (option) => {
-    handleCategoryChange({ target: { value: option.value } });
+    handleTypeChange({ target: { value: option.value } });
   };
 
   const customClassNames = {
@@ -20,19 +17,18 @@ const CategorySelect = ({ selectedCategory, categories, handleCategoryChange }) 
     option: "block w-full px-4 py-2 text-left text-orange-700 hover:bg-orange-100 transition-colors duration-200",
     dropdownContainer: "absolute z-10 w-full bg-white rounded-md shadow-lg mt-1 overflow-auto",
   };
-
   return (
     <Select
-      options={options}
+      options={option}
       onChange={handleChange}
-      label="Catégorie"
+      label="Type"
       icon={faList}
-      value={selectedCategory}
-      placeholder="Sélectionnez une catégorie"
+      value={selectedType}
+      placeholder="Sélectionnez un type"
       customClassNames={customClassNames}
-      maxHeight="200px" 
+      maxHeight="200px" // Vous pouvez ajuster cette valeur selon vos besoins
     />
   );
 };
 
-export default CategorySelect;
+export default TypeSelect;
