+++ frontend/src/pages/user/pages/components/selectedIngredient/getFullIngredient.jsx
